Deprecate \Psalm\getVendorDir() function
